refactor: split file, definition and routine
