+++ sa/CMakeLists.txt
@@ -328,6 +328,7 @@ set(METHOD_SOURCES
   ${METHOD_DIR}/method_connection_sr.cpp
   ${METHOD_DIR}/method_connection_java.cpp
   ${METHOD_DIR}/method_connection_pool.cpp
+  ${METHOD_DIR}/method_compile_def.cpp
   ${METHOD_DIR}/method_compile.cpp
   ${METHOD_DIR}/method_def.cpp
   ${METHOD_DIR}/method_error.cpp
